+++ wxc_sdk/common/__init__.py
@@ -1247,9 +1247,13 @@ class UsageType(str, Enum):
 
 
 class DirectLineCallerIdNameSelection(str, Enum):
-    #: When this option is selected, `customName` is to be shown for this auto attendant.
+    #: When this option is selected, `customName` is to be shown for this user.
     custom_name = 'CUSTOM_NAME'
-    #: When this option is selected, `name` is to be shown for this auto attendant.
+    #: When this option is selected, `firstName` and `lastName` are to be shown for this user.
+    first_name_last_name = 'FIRST_NAME_LAST_NAME'
+    #: When this option is selected, `lastName` and `firstName` are to be shown for this user.
+    last_name_first_name = 'LAST_NAME_FIRST_NAME'
+    #: When this option is selected, `displayName` is to be shown for this user.
     display_name = 'DISPLAY_NAME'